+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning][semver].
 
 ## [unreleased]
 
+## [0.1.0] - 2021-08-09
+
+### Added
+
+- add request interceptor `ConsoleLogRequestInterceptor`
+- add response interceptor `ConsoleLogResponseInterceptor`
+- some improvements
+
 ## [0.0.11] - 2021-06-22
 
 ### Fixed